# Implement OP_WRITE_RAW on the instructions buffer