--- a/ext/liquid_c/block.h
+++ b/ext/liquid_c/block.h
@@ -6,7 +6,6 @@
 typedef struct block_body {
     VALUE obj;
     vm_assembler_t code;
-    VALUE source; // hold a reference to the ruby object that OP_WRITE_RAW points to
     bool blank;
     int render_score;
     VALUE nodelist;