Implement `OP_WRITE_RAW` as an immediate instruction. The current implementation stores the size as a 24 byte unsigned integer after the instruction and then `size` number of bytes following it is the string. I also added a `OP_WRITE_RAW_SKIP` instruction that skips skips the string. This instruction is used in `BlockBody#remove_blank_strings` to replace `OP_WRITE_RAW` instruction. I'm not entirely satisfied in adding an extra instruction to handle this corner case. I've thought of two other ways to implement this:

1. Store an extra boolean flag to signal whether to skip the instruction or not. The upside to this is that we don't need the special `OP_WRITE_RAW_SKIP` instruction  but will require an extra 8 bytes.
2. Add a function to the `c_buffer` that allows deleting sections from the middle. This is probably the cleanest solution but will have performance overhead of `memmove`.

Also see #77.

# Benchmarks

Base branch:

```
              parse:    156.542  (± 3.2%) i/s -      1.575k in  10.074790s
             render:    222.599  (± 8.1%) i/s -      2.222k in  10.071418s
     parse & render:     85.181  (± 2.3%) i/s -    856.000  in  10.056397s
```

This branch:

```
              parse:    158.134  (± 8.9%) i/s -      1.573k in  10.057410s
             render:    230.214  (± 7.8%) i/s -      2.288k in  10.019494s
     parse & render:     87.223  (± 3.4%) i/s -    876.000  in  10.057597s
```