@dylanahsmith 2020-10-14
The benchmark results don't seem right.  I would expect parse time to be significantly slower due to having to copy the immediate argument for `OP_WRITE_RAW`.  Rendering also doesn't seem more efficient, since the string length needs to be decoded.  Did you get the benchmark results mixed up?

On master I got this for the benchmark

```
              parse:    156.458  (± 3.2%) i/s -      1.575k in  10.077279s
             render:    207.469  (± 5.3%) i/s -      2.080k in  10.058117s
     parse & render:     83.494  (± 3.6%) i/s -    840.000  in  10.070280s
```

and on this branch (rebased) I got 

```
              parse:    147.837  (± 4.1%) i/s -      1.484k in  10.053717s
             render:    203.815  (± 5.4%) i/s -      2.033k in  10.009205s
     parse & render:     80.104  (± 3.7%) i/s -    800.000  in  10.000488s
```

which seems inline with what I expected.

I was hoping to avoid slowing down rendering with this change.  I think we could have a more efficient OP_WRITE_RAW instruction for a byte sized length argument to make rendering more comparable to what we have on master.