@dylanahsmith 2020-10-16
Have you tried using a shorter OP_WRITE_RAW instruction to fix the performance regression?  How does that affect the benchmark results?